dev/sg: move RFC functions into package internal/rfc
